+++ cli/cli/command_framework/highlevel/file_system_path_arg/file_system_path_type.go
@@ -1,0 +1,10 @@
+package file_system_path_arg
+
+//go:generate go run github.com/dmarkham/enumer -type=FileSystemPathType -trimprefix=FileSystemPathType_ -transform=snake
+type FileSystemPathType int
+const (
+	FileSystemPathType_Filepath FileSystemPathType = iota
+	FileSystemPathType_Dirpath
+	FileSystemPathType_FilepathOrDirpath
+)
+